refactor: translate what-are-apps stories
- Move getCommunityStories into src/lib/utils/stories.ts and StoryData into @/lib/types, removing the app-directory reach-ins from src/data/tenYearStories.ts and the /stories page
- Split WhatAreAppsStories into a server component (resolves its three featured stories from the community-stories namespace, same pattern as sibling BrowseApps) and a client list with the expand state; drops the duplicated hard-coded story texts, localizes story copy and dates (date was hard-coded en-US), and translates the Read more/less labels via a new common key story-card-read-less
- Expose storyKey on resolved stories and use it as the React list key (fixes duplicate keys from the two "Nico" entries on /10years)
