+++ app/[locale]/10years/_components/types.ts
@@ -1,20 +1,5 @@
 import { StaticImageData } from "next/image"
 
-export type StoryData = {
-  /** Key in the "community-stories" namespace holding the translatable story copy */
-  storyKey: string
-  /** Story text verbatim as submitted (English for English submissions) */
-  storyOriginal: string
-  /** BCP-47 language code of the original submission (not necessarily a site locale) */
-  originalLocale: string
-  category: string
-  name: string
-  date: string
-  country: string
-  twitter: string
-  region: string
-}
-
 export type AdoptionCard = {
   image: StaticImageData
   title: string
