+++ app/[locale]/stories/_components/CommunityStories.tsx
@@ -94,9 +94,9 @@ const CommunityStories = ({ stories }: CommunityStoriesProps) => {
       {/* CSS multi-column gives true masonry packing with zero JS, so the
           server and client render identical markup (no hydration mismatch). */}
       <div className="gap-6 sm:columns-2 lg:columns-3 xl:columns-4">
-        {visibleStories.map((story, index) => (
+        {visibleStories.map((story) => (
           <StoryCard
-            key={`${story.name}-${index}`}
+            key={story.storyKey}
             story={story}
             expandable={false}
             showDate={false}
