+++ app/[locale]/stories/page.tsx
@@ -29,10 +29,10 @@ import { Section } from "@/components/ui/section"
 
 import { getStoriesData } from "@/lib/utils/md"
 import { getMetadata } from "@/lib/utils/metadata"
+import { getCommunityStories } from "@/lib/utils/stories"
 import { getRequiredNamespacesForPage } from "@/lib/utils/translations"
 import { getVideos } from "@/lib/utils/videos"
 
-import { getCommunityStories } from "../10years/_components/utils"
 import { getVideosByCategory } from "../videos/utils"
 
 import CommunityStories from "./_components/CommunityStories"
