+++ src/intl/en/common.json
@@ -510,6 +510,7 @@
   "show-more": "Show more",
   "show-less": "Show less",
   "story-card-read-more": "Read more",
+  "story-card-read-less": "Read less",
   "story-card-show-original": "Show original",
   "story-card-show-translation": "Show translation",
   "story-card-original-language": "Original language",
