+++ src/lib/types.ts
@@ -1421,7 +1421,25 @@ export type TimeLeftLabels = Record<
   TimeLeftLabel
 >
 
+/** Community story as authored in src/data/tenYearStories.ts */
+export type StoryData = {
+  /** Key in the "community-stories" namespace holding the translatable story copy */
+  storyKey: string
+  /** Story text verbatim as submitted (English for English submissions) */
+  storyOriginal: string
+  /** BCP-47 language code of the original submission (not necessarily a site locale) */
+  originalLocale: string
+  category: string
+  name: string
+  date: string
+  country: string
+  twitter: string
+  region: string
+}
+
+/** Community story resolved for rendering; see getCommunityStories */
 export type Story = {
+  storyKey: string
   name: string
   /** Story copy resolved to the viewer's locale (falls back to English) */
   story: string
